Update compilation modes.
This commit changes `(modes byte native)` to `(modes byte best)` in an
attempt to support platforms without the native compiler. The mode
`best` is supposed to default to `native`, and fallback to `byte` if
it is absent. Though, it is not clear whether dune will perform a
redundant compilation of the source code if `best` resolves to `byte`.