OCaml 5.2 support
This patch adds OCaml 5.2 support. The main change is to account for
the new continuation representation in OCaml 5.2, which represents a
continuation as a pair. Legacy support is preserved via the use of the
compile-time `MULTICONT52` variable. When toggled code for the new
continuation representation is included, otherwise this code is
excluded.

Resolves #6.