OCaml 5.2 support (#7)
* OCaml 5.2 support

This patch adds OCaml 5.2 support. The main change is to account for
the new continuation representation in OCaml 5.2, which represents a
continuation as a pair. Legacy support is preserved via the use of the
compile-time `MULTICONT52` variable. When toggled code for the new
continuation representation is included, otherwise this code is
excluded.

Resolves #6.

* Use ocaml-5.2.0+trunk in CI until the proper release is available.

* Follow the OCaml manual's FFI rules even if they may not be strictly required.